requirements.txt
flask

flask-Sqlalchemy
flask-WTF
flask-Login
wtforms[email]
flask-Migrate
flask-Mail

gunicorn

pytest
pytest-flask
pytest-cov